requirements.txt
Adafruit-Blinka==8.19.0
adafruit-circuitpython-bme680==3.4.13
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-register==1.9.16
adafruit-circuitpython-requests==1.14.0
adafruit-circuitpython-shtc3==1.1.12
adafruit-circuitpython-typing==1.9.4
Adafruit-PlatformDetect==3.46.0
Adafruit-PureIO==1.1.11
bme680==1.1.1
paho-mqtt==1.6.1
psutil==5.9.7
pyftdi==0.54.0
pyserial==3.5
python-dotenv==1.0.0
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
smbus==1.1.post2
sysv-ipc==1.1.0
typing-extensions==4.6.3
